LayoutTests/webanimations/keyframe-effect.html

<script src="../resources/js-test-pre.js"></script>
<script>

description("Check that the KeyframeEffect interface is defined and exposes a constructor.");

debug("The KeyframeEffect interface is defined.");
shouldBeDefined("KeyframeEffect");

debug("");
debug("The KeyframeEffect interface cannot be constructed without parameters.");
shouldThrow("new KeyframeEffect");

debug("");
debug("The KeyframeEffect interface constructor throws if the parameter has the wrong type.");
shouldThrow("(new KeyframeEffect(0))");
shouldThrow("(new KeyframeEffect({}))");
shouldThrow("(new KeyframeEffect([]))");
shouldThrow("(new KeyframeEffect(false))");
shouldThrow("(new KeyframeEffect('element'))");

debug("");
debug("The KeyframeEffect interface target is set to the provided Element.");
shouldBe("(new KeyframeEffect(document.body)).target", "document.body");

debug("");
debug("A KeyframeEffect object target is read-only.");
const effect = new KeyframeEffect(document.body);
effect.target = document.createElement("div");
shouldBe("effect.target", "document.body");

debug("");

</script>
<script src="../resources/js-test-post.js"></script>
