LayoutTests/webanimations/keyframe-effect-expected.txt

Check that the KeyframeEffect interface is defined and exposes a constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


The KeyframeEffect interface is defined.
PASS KeyframeEffect is defined.

The KeyframeEffect interface cannot be constructed without parameters.
PASS new KeyframeEffect threw exception TypeError: Not enough arguments.

The KeyframeEffect interface constructor throws if the parameter has the wrong type.
PASS (new KeyframeEffect(0)) threw exception TypeError: Argument 1 ('target') to the KeyframeEffect constructor must be an instance of Element.
PASS (new KeyframeEffect({})) threw exception TypeError: Argument 1 ('target') to the KeyframeEffect constructor must be an instance of Element.
PASS (new KeyframeEffect([])) threw exception TypeError: Argument 1 ('target') to the KeyframeEffect constructor must be an instance of Element.
PASS (new KeyframeEffect(false)) threw exception TypeError: Argument 1 ('target') to the KeyframeEffect constructor must be an instance of Element.
PASS (new KeyframeEffect('element')) threw exception TypeError: Argument 1 ('target') to the KeyframeEffect constructor must be an instance of Element.

The KeyframeEffect interface target is set to the provided Element.
PASS (new KeyframeEffect(document.body)).target is document.body

A KeyframeEffect object target is read-only.
PASS effect.target is document.body

PASS successfullyParsed is true

TEST COMPLETE
